break App and Links component out into separate files
