+++ src/styles.js
@@ -1,7 +1,7 @@
 const styles = ({ spacing }) => ({
   container: {
     // maxWidth: 600,
-    padding: spacing(3),
+    padding: spacing(2),
     fontFamily: "sans-serif",
     textAlign: "center",
     display: "flex",
